+++ changelog.md
@@ -12,6 +12,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) when it reaches version 1.0.
 - `ts.svg()`: Render geometries to an animated SVG (no dependencies required(!)).
 - `ts.svg()`: Added progress bar.
 - `ts.svg()`: Optionally show z,y,x-axes in top-left corner with `show_axes=True`.
+- `ts.reflect()`: Add reflection geometrical transform.
 ### Changed
 - `ts.svg()`: Remove `base64` keyword parameter. This functionality has been
   moved to the `_repr_markdown_` method.